pyproject.toml
[project]
name = "comfyui-firstordermm"
description = "ComfyUI-native nodes to run First Order Motion Model for Image Animation and its non-diffusion-based successors. [a/https://github.com/AliaksandrSiarohin/first-order-model](https://github.com/AliaksandrSiarohin/first-order-model)"
version = "1.0.5"
license = { file = "LICENSE" }
dependencies = ["numpy", "torch", "scipy", "pyyaml", "matplotlib", "einops", "timm"]

[project.urls]
Repository = "https://github.com/FuouM/ComfyUI-FirstOrderMM"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "fuoumarinas"
DisplayName = "ComfyUI-FirstOrderMM"
Icon = ""
